one user service was deleted
