--- eCommerce-Server/src/main/java/com/mate/hackathon/aug2018/ecommerce/service/UserServiceImp.java
@@ -1,35 +1,0 @@
-package com.mate.hackathon.aug2018.ecommerce.service;
-
-import com.mate.hackathon.aug2018.ecommerce.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-
-@Service
-public class UserServiceImp implements UserService {
-    @Autowired
-    private UserRepository repository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return repository.findByEmail(email)
-                .map(this::toUserDetails)
-                .orElseGet(() -> User.builder().disabled(true).build());
-    }
-
-    private UserDetails toUserDetails(com.mate.hackathon.aug2018.ecommerce.model.User user) {
-        return User.builder()
-                .username(user.getEmail())
-                .password(user.getPassword())
-                .authorities(Collections.emptyList())
-                .build();
-    }
-}
-
-
-
-
